* subversion/libsvn_subr/properties.c   (svn_property_kind): Quash a couple of warnings with casts (we can't change     the variable type because it is part of the public API).
git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1101060 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_subr/properties.c
@@ -72,14 +72,14 @@ svn_property_kind(int *prefix_len,
   if (strncmp(prop_name, SVN_PROP_WC_PREFIX, wc_prefix_len) == 0)
     {
       if (prefix_len)
-        *prefix_len = wc_prefix_len;
+        *prefix_len = (int) wc_prefix_len;
       return svn_prop_wc_kind;
     }
 
   if (strncmp(prop_name, SVN_PROP_ENTRY_PREFIX, entry_prefix_len) == 0)
     {
       if (prefix_len)
-        *prefix_len = entry_prefix_len;
+        *prefix_len = (int) entry_prefix_len;
       return svn_prop_entry_kind;
     }
